+++ packages/core/src/js/index.ts
@@ -4,6 +4,7 @@ export type {
   SdkInfo,
   Event,
   Exception,
+  SendFeedbackParams,
   SeverityLevel,
   StackFrame,
   Stacktrace,
@@ -16,6 +17,7 @@ export {
   addBreadcrumb,
   captureException,
   captureEvent,
+  captureFeedback,
   captureMessage,
   Scope,
   setContext,
@@ -41,6 +43,7 @@ export {
   setCurrentClient,
   addEventProcessor,
   metricsDefault as metrics,
+  lastEventId,
 } from '@sentry/core';
 
 export {